+++ web-components/aofl-drawer/documentation/routes/home-element/examples/no-animate/template.js
@@ -1,0 +1,8 @@
+(context, html) => html`
+<button class="button" on-click="${() => context.toggleDrawer()}">toggle</button>
+<aofl-drawer open$="${context.drawer}">
+  <p>
+    hello
+  </p>
+</aofl-drawer>
+`;