+++ web-components/aofl-drawer/documentation/routes/home-element/examples/side-panel/index.js
@@ -1,0 +1,5 @@
+context.drawer = 'false';
+
+context.toggleDrawer = () => {
+  context.drawer = context.drawer === 'false' ? 'true' : 'false';
+};