+++ web-components/aofl-drawer/documentation/routes/home-element/examples/side-panel/template.js
@@ -1,0 +1,8 @@
+(context, html) => html`
+<button class="button" on-click="${() => context.toggleDrawer()}">toggle</button>
+<aofl-drawer opening="slide-open" closing="slide-closed" open$="${context.drawer}">
+  <p>
+    side panel content
+  </p>
+</aofl-drawer>
+`;