+++ web-components/aofl-list-option/documentation/routes.config.js
@@ -1,13 +1,2 @@
 window.aofljsConfig = window.aofljsConfig || {};
-window.aofljsConfig.routesConfig = {
-  'routes': [
-    {
-      'resolve': () => import('./routes/home-element/index.js'),
-      'rotation': 'routes',
-      'path': '/',
-      'dynamic': false,
-      'title': 'AofL::Home',
-      'locale': ''
-    }
-  ]
-};
+window.aofljsConfig.routesConfig = {};